🎨 style(articles.php): rename Articles class to Article This commit renames the Articles class to Article to improve naming consistency and to better reflect the class's purpose.
🐛 fix(index.php): fix typo in Article class instantiation
✨ feat(index.php): add new article to articles array
Fixed a typo in the instantiation of the Article class, which was causing an error. Also, added a new article to the articles array, expanding the content of the page.

🚀 chore(index.php): add new objects to array
Added a new Commercial object and a new Vacancies object to the array of objects. This improves the content of the page by adding more variety and information.